[refactor] 프로젝트 구조 정리리
