--- .env
@@ -1,11 +1,0 @@
-# EXPRESS SERVER PORT
-PORT = 9999
-GOOGLE_CLIENT_ID=894080515848-nbimhst1528sj62qtl0sgspqn2eao0ho.apps.googleusercontent.com
-GOOGLE_CLIENT_SECRET=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
-JWT_SECRET=secret-key
-REDIRECT_URI=http://localhost:8080/loading
-DB_HOST=database-1.cru40cgcud6j.ap-northeast-2.rds.amazonaws.com
-DB_USER=lolsearch
-DB_PASSWORD=lolsearch
-DB_NAME=lol_search
-DB_PORT=3306
