+++ .gitignore
@@ -5,7 +5,7 @@ node_modules/
 .idea
 
 # .env 파일 제외
-#.env
+.env
 
 # 로그 파일들 제외
 *.log
