+++ bin/www
@@ -8,8 +8,7 @@ var app = require('../app');
 var debug = require('debug')('lol-search:server');
 var http = require('http');
 
-var dotenv = require('dotenv');
-dotenv.config();
+require('dotenv').config();
 
 /**
  * Get port from environment and store in Express.
