+++ controllers/commentController.js
@@ -1,4 +1,4 @@
-const auth = require('../token');
+const jwtUtil = require('../utils/jwt');
 const commentStore = require('../store/commentStore');
 
 const comment = {}
@@ -78,7 +78,7 @@ comment.createComment = async (req, res) => {
     const token = req.headers['authorization'];
     if (!token) return errorMessage(res, 401);
 
-    const authorization = auth.authenticateToken(token.split(' ')[1]);
+    const authorization = jwtUtil.verifyToken(token.split(' ')[1]);
     
     const result = await commentStore.createComment([authorization.id, postId, content]);
     /* 수정될 코드 */
@@ -110,7 +110,7 @@ comment.editComment = async (req, res) => {
     const token = req.headers['authorization'];
     if (!token) return errorMessage(res, 401);
 
-    const authorization = auth.authenticateToken(token.split(' ')[1]);
+    const authorization = jwtUtil.verifyToken(token.split(' ')[1]);
     if (userId != authorization.id) return errorMessage(res, 403);
     /* 수정될 코드 */
 
@@ -142,7 +142,7 @@ comment.delComment = async (req, res) => {
     const token = req.headers['authorization'];
     if (!token) return errorMessage(res, 401);
 
-    const authorization = auth.authenticateToken(token.split(' ')[1]);
+    const authorization = jwtUtil.verifyToken(token.split(' ')[1]);
     if (userId != authorization.id) return errorMessage(res, 403);
     /* 수정될 코드 */
