+++ routes/index.js
@@ -1,23 +1,6 @@
-var express = require('express');
-var router = express.Router();
+const express = require('express');
+const router = express.Router();
 
-/**
-* @swagger
-*   /:
-*     get:
-*       description: "default page API"
-*       responses:
-*         200:
-*           description: "success"
-*           content:
-*             application/json:
-*               schema:
-*                 type: string
-*                 example: "Welcome to Express"
-*/
-/**
-
-/* GET home page. */
 router.get('/', function(req, res, next) {
   res.render('index', { title: 'Express' });
 });
