+++ store/userStore.js
@@ -1,0 +1,19 @@
+const db = require('../config/db');
+
+class UserStore {
+  async findByGoogleId(googleId) {
+      const [rows] = await db.execute('SELECT * FROM user_table WHERE uuid = ?', [googleId]);
+      return rows[0];
+  }
+
+  async createUser(user) {
+    const { google_id, email, name } = user;
+    const [result] = await db.execute(
+        'INSERT INTO user_table (uuid, email, name) VALUES (?, ?, ?)',
+        [google_id, email, name]
+    );
+    return { id: result.insertId, ...user };
+  }
+}
+
+module.exports = new UserStore();
