+++ utils/jwt.js
@@ -6,5 +6,9 @@ exports.generateToken = (payload) => {
 };
 
 exports.verifyToken = (token) => {
-    return jwt.verify(token, process.env.JWT_SECRET);
+    try {
+        return jwt.verify(token, process.env.JWT_SECRET);
+    } catch (error) {
+        return error;
+    }
 };
